requirements.txt
flask==3.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==20.1.0 